[refactor] type 폴더 분리 - 명확한 관심사 분리를 위한 타입폴더 분리
