+++ src/services/auth/authService.ts
@@ -1,4 +1,4 @@
-import { OAuthDomainType } from './types';
+import { OAuthDomainType } from '../../types/auth/types';
 const { VITE_BASE_URI } = import.meta.env;
 // oauth 로그인 요청
 export const requestOAuth = (domain: OAuthDomainType): void => {
